backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.1
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.6
defusedxml==0.7.1
dj-rest-auth==2.1.3
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-ratings==0.3.7
django-rest-auth==0.9.5
django-sslserver==0.22
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
docker==4.4.3
gunicorn==20.0.4
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
optional-django==0.1.0
pathspec==0.8.1
phonenumbers==8.12.18
Pillow==8.2.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.7.2
python3-openid==3.2.0
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
stripe==2.57.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.3
websocket-client==0.57.0
wrapt==1.12.1
yapf==0.30.0